Add support for bypassing roll-prompt
I have added a new possible object variable for the RollForm called skipDialog.

This variable can only be used when rolling as part of a macro. When set to true, the object will skip the dialog prompt that is normally shown when rolling and it will immediately make the roll using the data provided when constructing the object.