Saved Rolls API update
Some crucial data under the Actor document was not being set up until after the user opens the sheet for the first time. This is bad.

I have mirrored the code from actor-sheet.js into actor.js so that these fields are initialized when the actor is initialized as opposed to only when the sheet is initialized.

Additionally, two new methods have been added to the Actor object:
• actor.savedRoll() which rolls a saved roll from the actor.
• actor.getSavedRoll() which returns the RollForm of the saved roll in case the user wants to make last minute edits.